Add support for AWS session_token
To be able to use AWS temporary credentials[1] the resource must
allow pass the session_token, together with the access key and
key id.

This is possibility is important in multiple scenarios, for instance
organizations that only allow usage of temporary AWS credentials,
or when using federated credentials. See [1] for more info.

In this commit we parse the `session_token` parameter, which
is passed to the s3client to build the credentials. If the parameter
is empty it will be ignore and the behaviour would be the same.

This commit also adds support to be able to pass the token when
running the tests, using the S3_TESTING_SESSION_TOKEN variable.

This allows testing the feature, but also run the tests in
organizations where only temporary are allowed.

A more specific test will be added in a separated commit.

To test this with some existing credentials, one can retrieve new
temporary credentials using awscli[2] as follows:

   S3_TESTING_ACCESS_KEY_ID=.... # as commented in the README
   S3_TESTING_SECRET_ACCESS_KEY=....
   S3_TESTING_BUCKET=...
   S3_VERSIONED_TESTING_BUCKET=...
   S3_TESTING_REGION=...
   S3_ENDPOINT=...

   # Get a new set of credentials
   eval $(
   AWS_SESSION_TOKEN= \
   AWS_ACCESS_KEY_ID="${S3_TESTING_ACCESS_KEY_ID}" \
   AWS_SECRET_ACCESS_KEY="${S3_TESTING_SECRET_ACCESS_KEY}" \
   aws sts get-session-token \
      --duration-seconds 900 \
      --output text \
      --query '[Credentials.AccessKeyId,Credentials.SecretAccessKey,Credentials.SessionToken]' \
      ${token:+--token-code "${token}"} | \
         awk '{ print "export S3_TESTING_ACCESS_KEY_ID=\"" $1 "\"\n" "export S3_TESTING_SECRET_ACCESS_KEY=\"" $2 "\"\n" "export S3_TESTING_SESSION_TOKEN=\"" $3 "\"" }'
   )

   # Check the new credentials
   AWS_SESSION_TOKEN="${S3_TESTING_SESSION_TOKEN}" \
   AWS_ACCESS_KEY_ID="${S3_TESTING_ACCESS_KEY_ID}" \
   AWS_SECRET_ACCESS_KEY="${S3_TESTING_SECRET_ACCESS_KEY}" \
      aws sts get-caller-identity

   # Run the tests
   docker build . -t s3-resource \
      --build-arg S3_TESTING_ACCESS_KEY_ID="${S3_TESTING_ACCESS_KEY_ID}" \
      --build-arg S3_TESTING_SECRET_ACCESS_KEY="${S3_TESTING_SECRET_ACCESS_KEY}" \
      --build-arg S3_TESTING_SESSION_TOKEN="${S3_TESTING_SESSION_TOKEN}" \
      --build-arg S3_TESTING_BUCKET="${S3_TESTING_BUCKET}" \
      --build-arg S3_VERSIONED_TESTING_BUCKET="${S3_VERSIONED_TESTING_BUCKET}" \
      --build-arg S3_TESTING_REGION="${S3_TESTING_REGION}" \
      --build-arg S3_ENDPOINT="${S3_ENDPOINT}"

[1] http://docs.aws.amazon.com/IAM/latest/UserGuide/id_credentials_temp.html
[2] https://aws.amazon.com/cli/