Add simple test for s3client with session token
In order to test the s3client with session token, we can use
the function sts.GetSessionToken[1] to generate some temporary
credentials. We add a small test that only checks if the bucket
can be accessed and the objects listed.

You do not need special permissions to call that function,
but can only be called with long-term AWS security credentials.

The test is disabled by default, but can be enabled by passing
the argument/env var TEST_SESSION_TOKEN

[1] http://docs.aws.amazon.com/STS/latest/APIReference/API_GetSessionToken.html

Signed-off-by: Caleb Washburn <cwashburn@pivotal.io>