Add ability to initialise non-existent resources
As described in https://github.com/concourse/s3-resource/issues/21, it
would be helpful if S3 resources could be configured to be available
even if they do not yet exist in S3. This commit adds parameters to
support that.

For S3 versioning, the `initial_version` parameter describes what the
version available should be named if the file does not exist in your
S3 bucket.

For filepath-based versioning, the `initial_path` parameter describes
the path to use if there are no regex matches in your S3 bucket.

A missing S3 resource will be empty by default, but you can specify
content for it. `initial_content_text` allows specifying text content,
or base64-encoded content can be provided to `initial_content_binary`.

Signed-off-by: Caleb Washburn <cwashburn@pivotal.io>