Revert "Merge pull request #96 from ghostsquad/feature/add-chain-provider"
This reverts commit b771dcda4c24b55ba720d6ed7f964425c3ce0337, reversing
changes made to 320096cafc335b60a6d6931802777289694f838b.

This breaks anonymous credentials use. We're seeing this error:

    error listing files: NoCredentialProviders: no valid providers in chain. Deprecated.

It seems that adding an empty static credentials provider to the chain
doesn't work. I think we'd need to use credentials.AnonymousCredentials
instead of the chain, but it's impossible to know whether to do that
because we already unconditionally append the EC2 role provider to the
chain. Maybe that should be made opt-in?

ref. concourse/s3-resource#96