Change prevent_file_overwrite to on_overwrite
We've realized that our use-case requires that the put succeed, but be a
no-op. To allow this I've changed the source field to on_overwrite,
which takes one of three options: "allow" (the default), "fail" (the
previous behavior) and "ignore" (the new behavior we need)

In implementing "ignore", I couldn't find a reasonable way to get the
VersionID of the existing file in S3. Due to this, I decided to disallow
the "ignore" setting when using "versioned_file", since this indicates
that the bucket is versioned (and VersionID is needed).

Signed-off-by: Matt Royal <mroyal@pivotal.io>