Perpare to withdraw the PR
- keep PR description in a file in the repo in case we resurrect this PR

REASON FOR WITHDRAWAL
The unpack_into param wouldn't solve our problem. Our problem was that the tar was archived with a root directory inside it, so even if we extracted it into a different path, it would stil have that directory in it. What would have been more useful is a way to pass the TAR's --strip-components=1 flag to strip the root dir when unpacking. Perhaps that's the PR we need. For now, shelving this.