upgrade aws-sdk to v2 and allow default AWS auth chain to be used
v2 is reaching EOL this year so we need to migrate off of v1.

This commit also changes the authentication behaviour of this resource.
We no longer prevent the use of EC2 IAM roles. If all you do is
configure a target bucket and file and provide none of the existing
authentication fields, we let the AWS sdk go through its default auth
chain which should result in the resource assuming the IAM attached to
the host worker.

Signed-off-by: Taylor Silva <dev@taydev.net>