Include versionId in generated public URLs
This commit restores the feature of displaying the correct link to a
versioned s3 file in the get task.  Currently, though the resource
creates resources correctly, the link on the Concourse interface does
not include the versionId URL parameter, leading to clicking on the link
always downloading the most recent version.

Signed-off-by: Chris Selzo <chris.selzo@broadcom.com>